API Reference

OpenXL.xl_parse — Function

xl_parse(x::AbstractString) -> XLWorkbook
xl_parse(x::Vector{UInt8}) -> XLWorkbook

Parse Excel file into XLWorkbook object.

Examples

julia> raw_xlsx = xl_sample_employee_xlsx()
48378-element Vector{UInt8}:
 0x50
 0x4b
    ⋮
 0x00

julia> xl_parse(raw_xlsx)
1-element XLWorkbook:
 1001x13 XLSheet("Employee")

Types

OpenXL.XLWorkbook — Type

XLWorkbook <: AbstractVector{XLSheet}

Represents an Excel workbook containing XLSheet.

Fields

sheets::Vector{XLSheet}: Workbook sheets.

Accessors

xl_names(x::XLWorkbook): Workbook sheet names.

See also: xl_parse.

OpenXL.AbstractXLSheet — Type

AbstractXLSheet <: AbstractArray{Any,2}

Abstract supertype for XLSheet and XLTable.

OpenXL.XLSheet — Type

XLSheet <: AbstractXLSheet

Sheet of the XLWorkbook. Supports indexing like a regular Matrix, as well as address indexing (e.g. A, A1, AB3 or range D:E, B1:C10, etc.).

The sheet slice will be converted into a separate XLTable.

Fields

name::String: Sheet name.

table::XLTable: Table representation.

Accessors

xl_name(x::XLSheet): Sheet name.

xl_nrow(x::XLTable): Number of rows.

xl_ncol(x::XLTable): Number of columns.

See also: xl_rowtable, xl_columntable, xl_print_sheet.

Examples

julia> xlsx = xl_parse(xl_sample_stock_xlsx())
1-element XLWorkbook:
 41x6 XLSheet("Stock")

julia> sheet = xlsx["Stock"]
41x6 XLSheet("Stock")
 Sheet │ A      B         C        D         E              F
───────┼─────────────────────────────────────────────────────────────────────
     1 │ name   price     h24      volume    mkt            sector
     2 │ MSFT   430.16    0.0007   11855456  3197000000000  Technology Serv…
     3 │ AAPL   189.98    -0.0005  36327000  2913000000000  Electronic Tech…
     4 │ NVDA   1064.69   0.0045   42948000  2662000000000  Electronic Tech…
     ⋮ │ ⋮      ⋮         ⋮        ⋮         ⋮              ⋮
    40 │ JNJ    146.97    0.0007   7.173e6   3.5371e11      Health Technolo…
    41 │ ORCL   122.91    -0.0003  5.984e6   3.3782e11      Technology Serv…

OpenXL.XLTable — Type

XLTable <: AbstractXLSheet

Table representation of the XLSheet object. Supports the same table operations as XLSheet.

Fields

data::Matrix: Table data.

Accessors

xl_nrow(x::XLTable): Number of rows.

xl_ncol(x::XLTable): Number of columns.

See also: xl_rowtable, xl_columntable, xl_print_sheet.

Examples

julia> xlsx = xl_parse(xl_sample_stock_xlsx())
1-element XLWorkbook:
 41x6 XLSheet("Stock")

julia> xlsx["Stock"]["A1:D25"]
25x4 XLTable
 Sheet │ A      B         C        D
───────┼───────────────────────────────────────
     1 │ name   price     h24      volume
     2 │ MSFT   430.16    0.0007   11855456
     3 │ AAPL   189.98    -0.0005  36327000
     4 │ NVDA   1064.69   0.0045   42948000
     ⋮ │ ⋮      ⋮         ⋮        ⋮
    24 │ NVDA   1064.69   0.0045   4.2948e7
    25 │ GOOG   176.33    -0.0006  1.1404e7

Methods

OpenXL.xl_rowtable — Function

xl_rowtable(sheet::AbstractXLSheet; kw...) -> Vector{NamedTuple}

Converts sheet rows to a Vector of NamedTuples.

Keyword arguments

alt_keys::Dict{String,String}: Alternative custom column headers.

header::Bool = false: Use first row elements as column headers.

Examples

julia> xlsx = xl_parse(xl_sample_stock_xlsx())
1-element XLWorkbook:
 41x6 XLSheet("Stock")

julia> xl_rowtable(xlsx["Stock"]["A1:C30"], header = true)
29-element Vector{NamedTuple{(:name, :price, :h24)}}:
 (name = "MSFT", price = "430.16", h24 = "0.0007")
 (name = "AAPL", price = "189.98", h24 = "-0.0005")
 (name = "NVDA", price = "1064.69", h24 = "0.0045")
 (name = "GOOG", price = "176.33", h24 = "-0.0006")
 ⋮
 (name = "LLY", price = 807.43, h24 = -0.0024)
 (name = "AVGO", price = 1407.84, h24 = 0.0036)

OpenXL.xl_columntable — Function

xl_columntable(sheet::AbstractXLSheet; kw...) -> Vector{NamedTuple}

Converts sheet columns to a Vector of NamedTuples.

Keyword arguments

alt_keys::Dict{String,String}: Alternative custom column headers.

header::Bool = false: Use first row elements as column headers.

Examples

julia> xlsx = xl_parse(xl_sample_stock_xlsx())
1-element XLWorkbook:
 41x6 XLSheet("Stock")

julia> alt_keys = Dict("A" => "Name", "B" => "Price", "C" => "H24");

julia> xl_columntable(xlsx["Stock"][2:end, 1:3]; alt_keys)
(
    Name = Any["MSFT", "AAPL"  …  "JNJ", "ORCL"]
    Price = Any[430.16, 189.98  …  146.97, 122.91],
    H24 = Any[0.0007, -0.0005  …  0.0007, -0.0003],
)

OpenXL.xl_print_sheet — Function

xl_print_sheet([io::IO], sheet::AbstractXLSheet; kw...)

Print a sheet as a table representation.

Keyword arguments

title::AbstractString = "Sheet": Table title in upper left corner.

header::Bool = false: Use first row elements as column headers.

max_len::Int = 16: Maximum length of an element in a cell.

compact::Bool = true: Omit rows and columns to save space.

Examples

julia> xlsx = xl_parse(xl_sample_employee_xlsx())
1-element XLWorkbook:
 1001x13 XLSheet("Employee")

julia> xl_print_sheet(xlsx["Employee"]; header = true)
 Sheet │ eeid    full_name        job_title         ⋯  country        city       exit_date  
───────┼───────────────────────────────────────────────────────────────────────────────────
     2 │ E02387  Emily Davis      Sr. Manger        ⋯  United States  Seattle    44485.0    
     3 │ E04105  Theodore Dinh    Technical Archi…  ⋯  China          Chongqing  nothing    
     4 │ E02572  Luna Sanders     Director          ⋯  United States  Chicago    nothing    
     5 │ E02832  Penelope Jordan  Computer System…  ⋯  United States  Chicago    nothing       
     ⋮ │ ⋮       ⋮                ⋮                 ⋯  ⋮              ⋮          ⋮             
  1000 │ E02521  Lily Nguyen      Sr. Analyst       ⋯  China          Chengdu    nothing    
  1001 │ E03545  Sofia Cheng      Vice President    ⋯  United States  Miami      nothing